[Refactor] 급여 도메인 사원, 관리자 구조 분리
