+++ hero-fe/src/views/payroll/me/BankAccountModal.vue
@@ -148,7 +148,7 @@
 <script setup lang="ts">
 import { ref, watch } from 'vue';
 import { usePayrollStore } from '@/stores/payrollStore';
-import type { BankAccount, MyPaySummary } from '@/types/payroll';
+import type { BankAccount, MyPaySummary } from '@/types/payroll/payroll.me';
 
 /**
  * props 설명
